prp-calculator/requirements.txt
# Core Dependencies
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.0

# PDF Generation
WeasyPrint==60.1
cairocffi==1.6.1

# Database
psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0